requirements.txt
Python 3.x
Gymnasium
PyTorch
Imageio
